feat(declarative) new _transform option in dc
This new config option (added to the declarative config file) allows importing basicauth credentials with and without hashed passwords.

The default value of `_transform` is `true`, which is backwards-compatible - in the past all imports always executed transformations.

However, *when exporting, it is always set to `false`*, as often the transformations are one-way (i.e. encryption of plain-text passwords).

Fixes #5649.