fix(declarative) use correct cache page when worker respawns
When worker respawns under DB-less mode, the `core_cache` is always uses
the first cache page. If the current page in use is the second page,
then the newly spawned worker will not be getting the correct view of
the configurations, causing undefined behaviors such as unexpected 404s
in router. This commit saves the `core_cache` page information inside
`shdict` so that new workers can always find the correct page to use.

fix #5799